Move code from `config` to `grpc`
Config should not contain any implementation itself, only serve as an
interface to functionality whose implementation is protocol-specific.
